Renamed module directory and associated testing directory
